backend/requirements.txt
# ── Core ──────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# ── Auth ─────────────────────────
PyJWT[cryptography]>=2.8.0   # [cryptography] extra required for RS256 support

# ── AI / LLM ─────────────────────
openai>=1.12.0               # DeepSeek & Groq use OpenAI-compatible API
httpx>=0.27.0                # async HTTP client for model calls
tiktoken>=0.6.0              # token counting for cost tracking

# ── Math Engine ──────────────────
sympy>=1.12

# ── OCR & Parsing ────────────────
pytesseract>=0.3.10
pix2tex>=0.1.1
Pillow>=10.2.0

# ── Database ─────────────────────
asyncpg>=0.29.0              # async Postgres driver
sqlalchemy[asyncio]>=2.0
supabase>=2.3.0
alembic>=1.13.1              # Database migrations
pgvector>=0.2.5              # Postgres vector extension support

# ── Cache ────────────────────────
redis[hiredis]>=5.0.0
numpy>=1.26.0                # vector operations for embeddings

# ── Background Workers ──────────
celery[redis]>=5.3.0

# ── Monitoring ───────────────────
sentry-sdk[fastapi]>=1.40.0
prometheus-client>=0.20.0
structlog>=24.1.0

# ── Testing ──────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx                        # TestClient transport

# ── Utilities ────────────────────
python-multipart>=0.0.6
